Removed the public/private switching for Explore Languages link
Testing on dev revealed that the generated url is dependent upon the last section that was visited (public/private).

Also, the explore section doesn't refresh the data when on the public tab and clicking the Explore link in the header. The end result is the theme switches to Workspaces but the page displays public dialects.